+++ packages/nextjs/components/tournament/play-moon.tsx
@@ -13,7 +13,7 @@ export const PlayMoon = () => {
   const { moonWallet } = useMoonWalletContext();
   const account = connectedAddress || moonWallet;
   const params = useParams<{ addr: string }>();
-  const chainId = 5;
+  const chainId = 80001;
   const [move, setMove] = useState(3);
   const [hash0, setHash0] = useState<`0x${string}`>("0x");
   const [hash1, setHash1] = useState<`0x${string}`>("0x");
