+++ packages/nextjs/components/tournament/withdraw.tsx
@@ -9,7 +9,7 @@ export const Withdraw = () => {
   const writeTx = useTransactor();
   // const { address: connectedAddress } = useAccount();
   const connectedAddress: string = useAccount()?.address ?? "";
-  const chainId = 5;
+  const chainId = 80001;
 
   const [withdrawn, setWithdrawn] = useState(false);
