+++ packages/nextjs/components/tournaments/item.tsx
@@ -6,7 +6,7 @@ import { ClockIcon, CurrencyDollarIcon, GiftIcon, TrophyIcon, UserCircleIcon } f
 import DeployedContracts from "~~/contracts/deployedContracts";
 
 export const Item = ({ tournament, player }: { tournament: string; player: boolean }) => {
-  const chainId = 5;
+  const chainId = 80001;
 
   const { data: tournamentData, isLoading: isLoading } = useContractRead({
     abi: DeployedContracts[chainId].Tournament.abi,
